Removed Dekstop packages Added WinUI specific Added Win10 WPF support
