--- Humbatt.UI.Toolkit.Desktop.MahApps/Humbatt.UI.Toolkit.Desktop.MahApps.csproj
@@ -1,32 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<TargetFrameworks>net8.0-windows;net9.0-windows</TargetFrameworks>
-		<RootNamespace>Humbatt.UI.Toolkit.Desktop</RootNamespace>
-		<Nullable>enable</Nullable>
-		<UseWPF>true</UseWPF>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Description>Humbatt.UI.Toolkit WPF extensions and helpers for .NET 8+ using MahApps</Description>
-		<PackageTags>WPF Controls</PackageTags>
-		<PackageReleaseNotes>Added support for .NET 9.0</PackageReleaseNotes>
-		<PackageReadmeFile>readme.md</PackageReadmeFile>
-		<Configurations>Debug;Release</Configurations>
-		<PackageReadmeFile>README.md</PackageReadmeFile>
-		<EnablePreviewMsixTooling>true</EnablePreviewMsixTooling>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<None Include="../README.md" Pack="true" PackagePath="\" />
-	</ItemGroup>
-	
-	<ItemGroup>
-		<PackageReference Include="MahApps.Metro" Version="2.4.10" />
-		<PackageReference Include="MahApps.Metro.IconPacks" Version="5.0.0" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\Humbatt.UI.Toolkit.Desktop\Humbatt.UI.Toolkit.Desktop.csproj" />
-	</ItemGroup>
-	
-</Project>
-
