--- Humbatt.UI.Toolkit.Desktop/Humbatt.UI.Toolkit.Desktop.csproj
@@ -1,95 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-	<PropertyGroup>
-		<TargetFrameworks>net8.0-windows7.0;net8.0-windows10.0.19041.0;net9.0-windows7.0;net9.0-windows10.0.19041.0;</TargetFrameworks>
-		<RootNamespace>Humbatt.UI.Toolkit.Desktop</RootNamespace>
-		<Description>WPF and WinUI UI controls, extensions and helpers for .NET 8.0+</Description>
-		<PackageTags>WPF WinUI</PackageTags>
-		<PackageReleaseNotes>Added support for .NET 9.0</PackageReleaseNotes>
-		<EnableDefaultCompileItems>false</EnableDefaultCompileItems>
-		<Configurations>Debug;Release</Configurations>
-		<EnableDefaultPageItems>false</EnableDefaultPageItems>
-		<PackageReadmeFile>README.md</PackageReadmeFile>
-		<EnablePreviewMsixTooling>true</EnablePreviewMsixTooling>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<None Include="../README.md" Pack="true" PackagePath="\" />
-	</ItemGroup>
-
-	<PropertyGroup Condition=" $(TargetFramework.Contains('-windows10')) ">
-		<TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion>
-		<RuntimeIdentifiers>win-x86;win-x64;win-arm64</RuntimeIdentifiers>
-		<UseWinUI>true</UseWinUI>
-		<DefineConstants>$(DefineConstants);WINUI</DefineConstants>
-		<WindowsSdkPackageVersion>10.0.19041.38</WindowsSdkPackageVersion>
-	</PropertyGroup>
-
-	<PropertyGroup Condition=" $(TargetFramework.Contains('-windows7.0')) ">
-		<UseWPF>true</UseWPF>
-		<DefineConstants>$(DefineConstants);WPF</DefineConstants>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<None Include="Models\RequiredValidationRule.wpf.cs" />
-		<Compile Include="**\*.shared.cs" />
-		<Compile Include="**\*.shared.*.cs" />
-	</ItemGroup>
-
-	<ItemGroup Condition=" $(TargetFramework.Contains('-windows10')) ">
-		<PackageReference Include="Microsoft.WindowsAppSDK" Version="1.6.240829007" />
-		<PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.26100.1" />
-		<PackageReference Include="Microsoft.Xaml.Behaviors.WinUI.Managed" Version="2.0.9" />
-		<PackageReference Include="CommunityToolkit.WinUI.UI.Controls.DataGrid" Version="7.1.2" />
-		<Compile Include="**\*.winui.cs" />
-		<Compile Include="**\*.winui.*.cs" />
-
-		<Page Include="..\WinUI\Generic.xaml" Link="Themes\Generic.xaml">
-			<Generator>MSBuild:Compile</Generator>
-		</Page>
-	</ItemGroup>
-
-	<ItemGroup Condition=" $(TargetFramework.Contains('-windows7.0')) ">
-		<PackageReference Include="Microsoft.Xaml.Behaviors.Wpf" Version="1.1.135" />
-		<Compile Include="**\*.wpf.cs" />
-		<Compile Include="**\*.wpf.*.cs" />
-
-
-		<Page Include="..\WPF\EditorThemes\Editors.xaml" Link="Editors\Themes\Editors.xaml">
-			<Generator>MSBuild:Compile</Generator>
-		</Page>
-
-		<Page Include="..\WPF\Generic.xaml" Link="Themes\Generic.xaml">
-			<Generator>MSBuild:Compile</Generator>
-		</Page>
-
-		<Page Include="Controls\FluidProgressBar.wpf.xaml">
-			<Generator>MSBuild:Compile</Generator>
-		</Page>
-
-	</ItemGroup>
-
-	<ItemGroup>
-		<Compile Remove="obj\**" />
-		<EmbeddedResource Remove="obj\**" />
-		<None Remove="obj\**" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<None Remove="Controls\FluidProgressBar.wpf.xaml" />
-		<None Remove="Resources\no-image-icon.png" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\Humbatt.UI.Toolkit.Core\Humbatt.UI.Toolkit.Core.csproj" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<Folder Include="Editors\Themes\" />
-		<Folder Include="Themes\" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<Resource Include="Resources\no-image-icon.png" />
-	</ItemGroup>
-
-</Project>
