--- Humbatt.UI.Toolkit.Desktop/Models/RequiredValidationRule.wpf.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Controls;
-
-namespace Humbatt.UI.Toolkit.Desktop.Models
-{
-	public class RequiredValidationRule : ValidationRule
-	{
-
-		public override ValidationResult Validate(object value, System.Globalization.CultureInfo cultureInfo)
-		{
-			var result = new ValidationResult(true, null);
-			if (value == null || ReferenceEquals(value, DBNull.Value))
-			{
-				result = new ValidationResult(false, "Required!");
-			}
-			return result;
-		}
-	}
-}
