+++ Humbatt.UI.Toolkit.WinUI/Converters/BaseConverter.winui.wpf.cs
@@ -10,7 +10,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Humbatt.UI.Toolkit.Desktop.Converters
+namespace Humbatt.UI.Toolkit.WinUI.Converters
 {
 	public abstract class BaseConverter : MarkupExtension
 	{
