+++ Humbatt.UI.Toolkit.WinUI/Converters/BooleanToReverseVisibilityConverter.winui.wpf.cs
@@ -5,51 +5,52 @@ using Microsoft.UI.Xaml;
 using System.Windows;
 using System.Windows.Data;
 #endif
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System;
 
-namespace Humbatt.UI.Toolkit.Desktop.Converters
+namespace Humbatt.UI.Toolkit.WinUI.Converters
 {
 
 #if WPF
 	[ValueConversion(typeof(bool), typeof(Visibility))]
 #endif
-    public class BooleanToVisibilityConverter : BaseConverter, IValueConverter
+	public class BooleanToReverseVisibilityConverter : BaseConverter, IValueConverter
 	{
-		public BooleanToVisibilityConverter()
+		public BooleanToReverseVisibilityConverter()
 		{
 
 		}
 
-		#region IValueConverter Members
 #if WINUI
 		public object Convert(object value, Type targetType, object parameter, string language)
 #elif WPF
 		public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
 #endif
+		
 		{
 			if (value == null)
-				return Visibility.Collapsed;
+				return Visibility.Visible;
 
 			if (!(value is bool))
-				return Visibility.Collapsed;
+				return Visibility.Visible;
 
 			var hide = Visibility.Collapsed;
 
-			if (parameter != null)
-			{
-				var pString = parameter.ToString();
+			//if (parameter != null)
+			//{
+			//	var pString = parameter.ToString();
 
-				if (pString.ToLower().Equals("hidden"))
-					hide = Visibility.Collapsed;
-			}
+			//	if (pString.Equals("collapsed", StringComparison.OrdinalIgnoreCase))
+			//		hide = Visibility.Collapsed;
+			//}
 
 			var bVal = (bool)value;
 
-			return (bVal == true) ? Visibility.Visible : hide;
+			return (bVal == true) ? hide : Visibility.Visible;
 		}
 
 #if WINUI
@@ -58,9 +59,7 @@ namespace Humbatt.UI.Toolkit.Desktop.Converters
 		public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
 #endif
 		{
-			throw new NotSupportedException();
+			throw new NotImplementedException();
 		}
-
-		#endregion
 	}
 }
