+++ Humbatt.UI.Toolkit.WinUI/Converters/BooleanToVisibilityConverter.winui.wpf.cs
@@ -5,52 +5,51 @@ using Microsoft.UI.Xaml;
 using System.Windows;
 using System.Windows.Data;
 #endif
-
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System;
 
-namespace Humbatt.UI.Toolkit.Desktop.Converters
+namespace Humbatt.UI.Toolkit.WinUI.Converters
 {
 
 #if WPF
 	[ValueConversion(typeof(bool), typeof(Visibility))]
 #endif
-	public class BooleanToReverseVisibilityConverter : BaseConverter, IValueConverter
+    public class BooleanToVisibilityConverter : BaseConverter, IValueConverter
 	{
-		public BooleanToReverseVisibilityConverter()
+		public BooleanToVisibilityConverter()
 		{
 
 		}
 
+		#region IValueConverter Members
 #if WINUI
 		public object Convert(object value, Type targetType, object parameter, string language)
 #elif WPF
 		public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
 #endif
-		
 		{
 			if (value == null)
-				return Visibility.Visible;
+				return Visibility.Collapsed;
 
 			if (!(value is bool))
-				return Visibility.Visible;
+				return Visibility.Collapsed;
 
 			var hide = Visibility.Collapsed;
 
-			//if (parameter != null)
-			//{
-			//	var pString = parameter.ToString();
+			if (parameter != null)
+			{
+				var pString = parameter.ToString();
 
-			//	if (pString.Equals("collapsed", StringComparison.OrdinalIgnoreCase))
-			//		hide = Visibility.Collapsed;
-			//}
+				if (pString.ToLower().Equals("hidden"))
+					hide = Visibility.Collapsed;
+			}
 
 			var bVal = (bool)value;
 
-			return (bVal == true) ? hide : Visibility.Visible;
+			return (bVal == true) ? Visibility.Visible : hide;
 		}
 
 #if WINUI
@@ -59,7 +58,9 @@ namespace Humbatt.UI.Toolkit.Desktop.Converters
 		public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
 #endif
 		{
-			throw new NotImplementedException();
+			throw new NotSupportedException();
 		}
+
+		#endregion
 	}
 }
