+++ Humbatt.UI.Toolkit.WinUI/Converters/DateToStringConverter.winui.wpf.cs
@@ -12,7 +12,7 @@ using System.Text;
 using System.Threading.Tasks;
 
 
-namespace Humbatt.UI.Toolkit.Desktop.Converters
+namespace Humbatt.UI.Toolkit.WinUI.Converters
 {
 
 #if WPF
