+++ Humbatt.UI.Toolkit.WinUI/Converters/EnumMatchToBooleanConverter.winui.wpf.cs
@@ -12,7 +12,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Humbatt.UI.Toolkit.Desktop.Converters
+namespace Humbatt.UI.Toolkit.WinUI.Converters
 {
 	/// <summary>
 	/// Converter for switching converting between an enum and boolean for use in a Radio button group
