+++ Humbatt.UI.Toolkit.WinUI/Themes/Generic.xaml
@@ -1,0 +1,19 @@
+<ResourceDictionary
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local2="using:Humbatt.UI.Toolkit.WinUI.Controls">
+
+    <Style TargetType="local2:SearchableListControl">
+        <Setter Property="Template">
+            <Setter.Value>
+                <ControlTemplate TargetType="local2:SearchableListControl">
+                    <Border
+                        Background="{TemplateBinding Background}"
+                        BorderBrush="{TemplateBinding BorderBrush}"
+                        BorderThickness="{TemplateBinding BorderThickness}">
+                    </Border>
+                </ControlTemplate>
+            </Setter.Value>
+        </Setter>
+    </Style>
+</ResourceDictionary>
