+++ WpfSample/WpfSample.csproj
@@ -10,11 +10,6 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\Humbatt.UI.Toolkit.Desktop.MahApps\Humbatt.UI.Toolkit.Desktop.MahApps.csproj" />
-		<ProjectReference Include="..\Humbatt.UI.Toolkit.Desktop\Humbatt.UI.Toolkit.Desktop.csproj" />
-	</ItemGroup>
-
-	<ItemGroup>
 		<PackageReference Include="DSoft.System.Mvvm" Version="3.4.2406.251" />
 	</ItemGroup>
